+++ test/mocks/factories/chainStatusFactory.ts
@@ -2,7 +2,7 @@ import {BigNumber} from 'ethers';
 import {Factory} from 'rosie';
 
 import {ChainStatus} from '../../../src/types/ChainStatus';
-import {ChainStatusWithAddress, ChainsStatuses} from '../../../src/types/ChainStatus';
+import {ChainStatusWithAddress} from '../../../src/types/ChainStatus';
 
 export const chainStatusFactory = Factory.define<ChainStatus>('LeavesAndFeeds')
   .attr('blockNumber', BigNumber.from(132153))
@@ -21,9 +21,3 @@ export const chainStatusWithAddressFactory = Factory.define<ChainStatusWithAddre
   .attr('chainAddress', '0xabc123')
   .attr('chainId', 'bsc')
   .attr('chainStatus', chainStatusFactory.build());
-
-export const multiChainStatusesFactory = Factory.define<ChainsStatuses>('MultiChainStatuses')
-  .attr('nextLeader', '0xaebd')
-  .attr('validators', ['0xaebd', '0xabctest'])
-  .attr('chainsIdsReadyForBlock', ['bsc'])
-  .attr('chainsStatuses', [chainStatusWithAddressFactory.build()]);