+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- cache masterchain data required to run consensus
+
+### Changed
+- set default timeout to 5s for `chain.getStatus()`
+
+### Removed
+- remove deprecated leader selection
 
 ## [7.6.5] - 2023-01-05
 ### Fixed